+++ hibernate-based-migration-example/src/main/java/dev/bodewig/hibernate_based_migration_example/Fruit.java
@@ -1,0 +1,12 @@
+package dev.bodewig.hibernate_based_migration_example;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+
+@Entity
+public class Fruit {
+	@Id
+	public String name;
+	public int weight;
+	public String colorRgb;
+}
