+++ hibernate-based-migration-example/src/main/resources/db/V0_hibernate.cfg.xml
