+++ hibernate-based-migration-example/src/main/resources/db/V1_hibernate.cfg.xml
