--- java-based-migration-example/pom.xml
@@ -1,171 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>dev.bodewig.java-based-migration</groupId>
-        <artifactId>java-based-migration-parent</artifactId>
-        <version>1.0.0-SNAPSHOT</version>
-    </parent>
-    <artifactId>java-based-migration-example</artifactId>
-    <version>2.0.0</version>
-    <packaging>jar</packaging>
-    <dependencies>
-        <dependency>
-            <groupId>com.h2database</groupId>
-            <artifactId>h2</artifactId>
-            <version>2.1.214</version>
-        </dependency>
-        <dependency>
-            <groupId>org.flywaydb</groupId>
-            <artifactId>flyway-core</artifactId>
-            <version>9.20.0</version>
-        </dependency>
-        <dependency>
-            <groupId>org.hibernate.orm</groupId>
-            <artifactId>hibernate-core</artifactId>
-            <version>6.2.6.Final</version>
-        </dependency>
-        <dependency>
-            <groupId>dev.bodewig.db2ascii</groupId>
-            <artifactId>db2ascii</artifactId>
-            <version>1.0.0</version>
-        </dependency>
-        <dependency>
-            <groupId>dev.bodewig.java-based-migration</groupId>
-            <artifactId>hibernate-based-migration</artifactId>
-            <version>1.0.0-SNAPSHOT</version>
-        </dependency>
-    </dependencies>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-clean-plugin</artifactId>
-                <version>3.2.0</version>
-                <executions>
-                    <execution>
-                        <id>clean-migration</id>
-                        <goals>
-                            <goal>clean</goal>
-                        </goals>
-                        <phase>clean</phase>
-                        <configuration>
-                            <excludeDefaultDirectories>true</excludeDefaultDirectories>
-                            <filesets>
-                                <fileset>
-                                    <directory>${project.basedir}/src/migration/${project.verion}</directory>
-                                </fileset>
-                            </filesets>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>dev.bodewig.java-based-migration</groupId>
-                <artifactId>java-based-migration-plugin</artifactId>
-                <version>1.0.0-SNAPSHOT</version>
-                <executions>
-                    <execution>
-                        <id>freeze</id>
-                        <goals>
-                            <goal>freeze</goal>
-                        </goals>
-                        <phase>validate</phase>
-                        <configuration>
-                            <freezeVersion>${project.version}</freezeVersion>
-                            <frozenDir>${project.basedir}/src/migration/</frozenDir>
-                            <persistenceClassesFileList>
-                                <file>${project.build.sourceDirectory}/my/db/classes/Person.java</file>
-                                <file>${project.build.sourceDirectory}/my/db/classes/pkg/</file>
-                            </persistenceClassesFileList>
-                            <persistenceClassesGlobList>
-                                <glob>src/main/java/my/db/classes/C*.java</glob>
-                                <glob>
-                                    <basePath>${project.build.sourceDirectory}</basePath>
-                                    <pattern>Bi*.java</pattern>
-                                </glob>
-                            </persistenceClassesGlobList>
-                            <persistenceResourcesFileList>
-                                <file>${project.basedir}/src/main/resources/myproperties.yaml</file>
-                            </persistenceResourcesFileList>
-                            <persistenceResourcesGlobList>
-                                <glob>
-                                    <basePath>${project.basedir}/src/main/resources/</basePath>
-                                    <pattern>*.yaml</pattern>
-                                </glob>
-                            </persistenceResourcesGlobList>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>thaw</id>
-                        <goals>
-                            <goal>thaw</goal>
-                        </goals>
-                        <phase>initialize</phase>
-                        <configuration>
-                            <frozenDir>${project.basedir}/src/migration/</frozenDir>
-                            <versionRange>[1.0.0,${project.version}]</versionRange>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>exec-maven-plugin</artifactId>
-                <version>3.1.0</version>
-                <executions>
-                    <execution>
-                        <id>pure-java</id>
-                        <goals>
-                            <goal>java</goal>
-                        </goals>
-                        <phase>test</phase>
-                        <configuration>
-                            <mainClass>my.db.migration.example.PureJavaExample</mainClass>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>flyway</id>
-                        <goals>
-                            <goal>java</goal>
-                        </goals>
-                        <phase>test</phase>
-                        <configuration>
-                            <mainClass>my.db.migration.example.FlywayExample</mainClass>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>hibernate</id>
-                        <goals>
-                            <goal>java</goal>
-                        </goals>
-                        <phase>test</phase>
-                        <configuration>
-                            <mainClass>my.db.migration.example.HibernateExample</mainClass>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>hibernate-stateless</id>
-                        <goals>
-                            <goal>java</goal>
-                        </goals>
-                        <phase>test</phase>
-                        <configuration>
-                            <mainClass>my.db.migration.example.HibernateStatelessExample</mainClass>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>hibernate-migration</id>
-                        <goals>
-                            <goal>java</goal>
-                        </goals>
-                        <phase>test</phase>
-                        <configuration>
-                            <mainClass>my.db.migration.example.HibernateMigrationExample</mainClass>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-</project>
