--- java-based-migration-example/src/main/java/my/db/classes/Bicycle.java
@@ -1,5 +1,0 @@
-package my.db.classes;
-
-public class Bicycle {
-	public String id;
-}
