--- java-based-migration-example/src/main/java/my/db/classes/Bike.java
@@ -1,5 +1,0 @@
-package my.db.classes;
-
-public class Bike {
-	public String id;
-}
