--- java-based-migration-example/src/main/java/my/db/classes/Car.java
@@ -1,6 +1,0 @@
-package my.db.classes;
-
-public class Car {
-	public String brand;
-	public String model;
-}
