--- java-based-migration-example/src/main/java/my/db/classes/Person.java
@@ -1,13 +1,0 @@
-package my.db.classes;
-
-import my.db.classes.pkg.food.Fruit;
-
-public class Person {
-	public String name;
-	public int age;
-	public Fruit favoriteFruit;
-
-	public Fruit getFavoriteFruit() {
-		return this.favoriteFruit;
-	}
-}
