--- java-based-migration-example/src/main/java/my/db/classes/pkg/food/Fruit.java
@@ -1,12 +1,0 @@
-package my.db.classes.pkg.food;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-
-@Entity
-public class Fruit {
-	@Id
-	public String name;
-	public int weight;
-	public String colorRgb;
-}
