--- java-based-migration-example/src/main/resources/__META-INF/persistence.xml
@@ -1,34 +1,0 @@
-<persistence
-  xmlns="http://java.sun.com/xml/ns/persistence"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd"
-  version="2.0"
->
-	<persistence-unit name="hibernate-old">
-		<provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-		<class>my.db._1_0_0.classes.pkg.food.Fruit</class>
-		<exclude-unlisted-classes>true</exclude-unlisted-classes>
-		<properties>
-			<property name="hibernate.dialect" value="org.hibernate.dialect.H2Dialect" />
-			<property name="hibernate.hbm2ddl.auto" value="create-drop" />
-			<property name="jakarta.persistence.jdbc.url" value="jdbc:h2:mem:hibernate" />
-			<property name="jakarta.persistence.jdbc.user" value="" />
-			<property name="jakarta.persistence.jdbc.password" value="" />
-			<property name="hibernate.connection.autocommit" value="true" />
-		</properties>
-	</persistence-unit>
-	
-	<persistence-unit name="hibernate-new">
-		<provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-		<class>my.db._2_0_0.classes.pkg.food.Fruit</class>
-		<exclude-unlisted-classes>true</exclude-unlisted-classes>
-		<properties>
-			<property name="hibernate.dialect" value="org.hibernate.dialect.H2Dialect" />
-			<property name="hibernate.hbm2ddl.auto" value="update" />
-			<property name="jakarta.persistence.jdbc.url" value="jdbc:h2:mem:hibernate" />
-			<property name="jakarta.persistence.jdbc.user" value="" />
-			<property name="jakarta.persistence.jdbc.password" value="" />
-			<property name="hibernate.connection.autocommit" value="true" />
-		</properties>
-	</persistence-unit>
-</persistence>
