--- java-based-migration-example/src/main/resources/myproperties.json
@@ -1,7 +1,0 @@
-{
-  "my": {
-    "new": {
-      "properties": [ ]
-    }
-  }
-}
