--- java-based-migration-example/src/main/resources/myproperties.yaml
@@ -1,4 +1,0 @@
----
-my:
-  old:
-    properties: []
