--- java-based-migration-example/src/migration/1.0.0/java/my/db/_1_0_0/classes/Bicycle.java
@@ -1,4 +1,0 @@
-package my.db._1_0_0.classes;
-public class Bicycle {
-    public String id;
-}
