--- java-based-migration-example/src/migration/1.0.0/java/my/db/_1_0_0/classes/Bike.java
@@ -1,4 +1,0 @@
-package my.db._1_0_0.classes;
-public class Bike {
-    public String id;
-}
