--- java-based-migration-example/src/migration/2.0.0/java/my/db/_2_0_0/classes/Bicycle.java
@@ -1,4 +1,0 @@
-package my.db._2_0_0.classes;
-public class Bicycle {
-    public String id;
-}
