--- java-based-migration-example/src/migration/2.0.0/java/my/db/_2_0_0/classes/Bike.java
@@ -1,4 +1,0 @@
-package my.db._2_0_0.classes;
-public class Bike {
-    public String id;
-}
