--- java-based-migration-example/src/migration/2.0.0/java/my/db/_2_0_0/classes/pkg/food/Fruit.java
@@ -1,12 +1,0 @@
-package my.db._2_0_0.classes.pkg.food;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-@Entity
-public class Fruit {
-    @Id
-    public String name;
-
-    public int weight;
-
-    public String colorRgb;
-}
