--- java-based-migration-example/src/migration/2.0.0/resources/myproperties.yaml
@@ -1,4 +1,0 @@
----
-my:
-  old:
-    properties: []
