+++ pom.xml
@@ -1,13 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>dev.bodewig.java-based-migration</groupId>
-    <artifactId>java-based-migration-parent</artifactId>
+    <groupId>dev.bodewig.hibernate-based-migration</groupId>
+    <artifactId>hibernate-based-migration-parent</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
     <modules>
-        <module>java-based-migration-plugin</module>
-        <module>java-based-migration-example</module>
+        <module>hibernate-based-migration</module>
+        <module>hibernate-based-migration-plugin</module>
+        <module>hibernate-based-migration-example</module>
     </modules>
     <properties>
         <java.version>17</java.version>
